+++ Changelog.md
@@ -4,7 +4,9 @@
 
 ### ISO Registry API Change
 
-- Changes in the ``registry.items()`` method API. This method now accepts an empty/missing ``region_codes`` argument to retrieve the full registry. Please see the [ISO Registry documentation](https://peopledoc.github.io/workalendar/iso-registry.html) for extensive usage docs (#403, #375).
+- Changes in the ``registry.items()`` method API.
+  - This method is aliased to ``get_calendars()``. In a near release, the ``items()`` method will change its purpose.
+  - The ``get_calendars()`` method accepts an empty/missing ``region_codes`` argument to retrieve the full registry. Please see the [ISO Registry documentation](https://peopledoc.github.io/workalendar/iso-registry.html) for extensive usage docs (#403, #375).
 
 ## v7.2.0 (2019-12-06)